added : font

+++ 2025_winapi_framework_22/Player.cpp
@@ -21,7 +21,7 @@
 
 Player::Player() : m_isDead(false), m_life(3), m_powerLevel(0), m_isInvincible(false),
 m_projCooldown(0.f), m_bombCnt(0), m_invincibleTime(0.f), m_bombDurationTimer(0.f),
-m_amountDmg(0),
+m_amountDmg(0.f),
 col(nullptr), fsm(nullptr), m_health(nullptr),  m_proj(nullptr)
 {
 	m_pTex = GET_SINGLE(ResourceManager)->GetTexture(L"Jiwoo");
@@ -141,7 +141,7 @@ void Player::CreateProjectile()
 		start_angle_deg = -(num_proj - 1) * ANGLE_STEP / 2.f;
 	}
 
-	int baseDmg = 10;
+	float baseDmg = 10.f;
 	int totalDmg = baseDmg + m_amountDmg;
 
 	for (int i = 0; i < num_proj; ++i)
@@ -253,7 +253,7 @@ void Player::InvokeBomb() {
 
 void Player::GainPower(int _amount) {
 	m_powerLevel += _amount;
-	m_amountDmg += 1;
+	m_amountDmg += 1.f;
 
 	m_powerLevel = std::min(m_powerLevel, MAX_POWER);
 	std::cout << "Power Level: " << m_powerLevel << std::endl;
